+++ src/mlpack/core/data/one_hot_encoding.hpp
@@ -14,6 +14,7 @@
 #define MLPACK_CORE_DATA_ONE_HOT_ENCODING_HPP
 
 #include <mlpack/prereqs.hpp>
+#include <mlpack/core.hpp>
 
 namespace mlpack {
 namespace data {
@@ -28,9 +29,38 @@ namespace data {
  * @param labelsIn Input labels of arbitrary datatype.
  * @param output Binary matrix.
  */
-template<typename eT, typename RowType>
+template<typename RowType, typename MatType>
 void OneHotEncoding(const RowType& labelsIn,
+                    MatType& output);
+
+/**
+ * Overloaded function for the above function, which takes a matrix as input
+ * and also a vector of indices to encode and outputs a matrix.
+ * Indices represent the IDs of the dimensions to be one-hot encoded.
+ *
+ * @param input Input dataset to be encoded.
+ * @param indices Index of rows to be encoded.
+ * @param output Encoded matrix.
+ */
+template<typename eT>
+void OneHotEncoding(const arma::Mat<eT>& input,
+                    const arma::Col<size_t>& indices,
                     arma::Mat<eT>& output);
+
+/**
+ * Overloaded function for the above function, which takes a matrix as input
+ * and also a DatasetInfo object and outputs a matrix.
+ * This function encodes all the dimensions marked `Datatype::categorical`
+ * in the data::DatasetInfo.
+ *
+ * @param input Input dataset to be encoded.
+ * @param output Encoded matrix.
+ * @param datasetInfo DatasetInfo object that has information about data.
+ */
+template<typename eT>
+void OneHotEncoding(const arma::Mat<eT>& input,
+                    arma::Mat<eT>& output,
+                    const data::DatasetInfo& datasetInfo);
 
 } // namespace data
 } // namespace mlpack