+++ src/mlpack/methods/ann/layer/layer.hpp
@@ -67,6 +67,7 @@
 #include "sequential.hpp"
 #include "softshrink.hpp"
 #include "softmax.hpp"
+#include "spatial_dropout.hpp"
 #include "subview.hpp"
 #include "transposed_convolution.hpp"
 #include "virtual_batch_norm.hpp"