+++ src/mlpack/methods/ann/layer/layer_types.hpp
@@ -48,6 +48,7 @@
 #include <mlpack/methods/ann/layer/reparametrization.hpp>
 #include <mlpack/methods/ann/layer/select.hpp>
 #include <mlpack/methods/ann/layer/softmax.hpp>
+#include <mlpack/methods/ann/layer/spatial_dropout.hpp>
 #include <mlpack/methods/ann/layer/subview.hpp>
 #include <mlpack/methods/ann/layer/virtual_batch_norm.hpp>
 #include <mlpack/methods/ann/layer/hardshrink.hpp>
@@ -291,6 +292,7 @@ using LayerTypes = boost::variant<
     PReLU<arma::mat, arma::mat>*,
     PixelShuffle<arma::mat, arma::mat>*,
     Softmax<arma::mat, arma::mat>*,
+    SpatialDropout<arma::mat, arma::mat>*,
     TransposedConvolution<NaiveConvolution<ValidConvolution>,
             NaiveConvolution<ValidConvolution>,
             NaiveConvolution<ValidConvolution>, arma::mat, arma::mat>*,